ChatMessage/requirements.txt
boto3==1.26.21  # or latest stable version
google-generativeai